requirements.txt
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
future==0.18.2
google-auth==1.15.0
google-auth-oauthlib==0.4.1
gspread==3.6.0
gspread-pandas==2.2.3
idna==2.9
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
jupyter-client==6.1.3
jupyter-core==4.6.3
langdetect==1.0.8
lxml==4.5.1
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
python-dateutil==2.8.1
python-twitter==3.5
pytz==2020.1
pyzmq==19.0.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
